+++ app/src/main/java/com/hhp227/knu_minigroup/fragment/GroupFragment.java
@@ -406,7 +406,9 @@ public class GroupFragment extends Fragment {
                     } catch (Exception e) {
                         Log.e(TAG, e.getMessage());
                     } finally {
-                        requireActivity().getWindow().clearFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE);
+                        if (getActivity() != null) {
+                            getActivity().getWindow().clearFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE);
+                        }
                     }
                 } else {
                     if (dataSnapshot.hasChildren()) {
@@ -415,8 +417,11 @@ public class GroupFragment extends Fragment {
 
                             fetchDataTaskFromFirebase(databaseReference.child(snapshot.getKey()), true);
                         }
-                    } else
-                        requireActivity().getWindow().clearFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE);
+                    } else {
+                        if (getActivity() != null) {
+                            getActivity().getWindow().clearFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE);
+                        }
+                    }
                 }
             }
